+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+- Export `semanticDispatch` 
+
 ## [1.1.0] - 2026-01-23
 
 ### Breaking changes